+++ internal/publish/connect_cloud/content_request_base.go
@@ -12,25 +12,25 @@ import (
 
 func getCloudContentType(contentType config.ContentType) (types.ContentType, error) {
 	switch contentType {
-	case config.ContentTypeHTML, config.ContentTypeJupyterNotebook, config.ContentTypeJupyterVoila:
+	case config.ContentTypeJupyterNotebook:
 		return types.ContentTypeJupyter, nil
 	case config.ContentTypePythonBokeh:
 		return types.ContentTypeBokeh, nil
 	case config.ContentTypePythonDash:
 		return types.ContentTypeDash, nil
-	case config.ContentTypePythonShiny, config.ContentTypeRShiny, config.ContentTypeQuartoShiny, config.ContentTypeRMarkdownShiny:
+	case config.ContentTypePythonShiny, config.ContentTypeRShiny:
 		return types.ContentTypeShiny, nil
 	case config.ContentTypePythonStreamlit:
 		return types.ContentTypeStreamlit, nil
-	case config.ContentTypeQuartoDeprecated, config.ContentTypeQuarto:
+	case config.ContentTypeQuartoDeprecated, config.ContentTypeQuarto, config.ContentTypeHTML:
 		return types.ContentTypeQuarto, nil
 	case config.ContentTypeRMarkdown:
 		return types.ContentTypeRMarkdown, nil
 	}
-	return "", fmt.Errorf("unsupported content type: %s", contentType)
+	return "", fmt.Errorf("content type '%s' is not supported by Connect Cloud", contentType)
 }
 
-func (c *ServerPublisher) getContentRequestBase() (*types.ContentRequestBase, error) {
+func (c *ServerPublisher) getContentRequestBase(isFirstDeploy bool) (*types.ContentRequestBase, error) {
 
 	// Extract config details for the request
 	title := c.Config.Title
@@ -100,20 +100,26 @@ func (c *ServerPublisher) getContentRequestBase() (*types.ContentRequestBase, er
 		vanityName = cloudCfg.VanityName
 	}
 
-	return &types.ContentRequestBase{
+	revision := types.RequestRevision{
+		SourceType:    "bundle",
+		RVersion:      rVersion,
+		PythonVersion: pythonVersion,
+		ContentType:   cloudContentType,
+		AppMode:       appMode,
+		PrimaryFile:   c.Config.Entrypoint,
+	}
+	base := &types.ContentRequestBase{
 		Title:       title,
 		Description: c.Config.Description,
-		NextRevision: types.NextRevision{
-			SourceType:    "bundle",
-			RVersion:      rVersion,
-			PythonVersion: pythonVersion,
-			ContentType:   cloudContentType,
-			AppMode:       appMode,
-			PrimaryFile:   c.Config.Entrypoint,
-		},
-		Access:     access,
-		AppMode:    types.AppModeFromType(c.Config.Type),
-		Secrets:    secrets,
-		VanityName: vanityName,
-	}, nil
+		Access:      access,
+		AppMode:     types.AppModeFromType(c.Config.Type),
+		Secrets:     secrets,
+		VanityName:  vanityName,
+	}
+	if isFirstDeploy {
+		base.NextRevision = &revision
+	} else {
+		base.RevisionOverrides = &revision
+	}
+	return base, nil
 }