+++ internal/util/files_test.go
@@ -7,9 +7,10 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/posit-dev/publisher/internal/util/utiltest"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/posit-dev/publisher/internal/util/utiltest"
 )
 
 // Copyright (C) 2023 by Posit Software, PBC.
@@ -20,39 +21,6 @@ type FilesSuite struct {
 
 func TestFilesSuite(t *testing.T) {
 	suite.Run(t, new(FilesSuite))
-}
-
-func (s *FilesSuite) TestDirFromPathDir() {
-	fs := afero.NewMemMapFs()
-	err := fs.MkdirAll("/my/dir", 0700)
-	s.Nil(err)
-
-	path := NewPath("/my/dir", fs)
-	dir, err := DirFromPath(path)
-	s.Nil(err)
-	s.Equal(path, dir)
-}
-
-func (s *FilesSuite) TestDirFromPathDirErr() {
-	fs := afero.NewMemMapFs()
-	path := NewPath("/nonexistent", fs)
-	dir, err := DirFromPath(path)
-	s.NotNil(err)
-	s.ErrorIs(err, os.ErrNotExist)
-	s.Equal(Path{}, dir)
-}
-
-func (s *FilesSuite) TestDirFromPathFile() {
-	fs := afero.NewMemMapFs()
-	err := fs.MkdirAll("/my/dir", 0700)
-	s.Nil(err)
-	err = afero.WriteFile(fs, "/my/dir/app.py", nil, 0600)
-	s.Nil(err)
-
-	path := NewPath("/my/dir/app.py", fs)
-	dir, err := DirFromPath(path)
-	s.Nil(err)
-	s.Equal(path.Dir(), dir)
 }
 
 func (s *FilesSuite) TestChdir() {